+++ api/image_pull_test.go
@@ -22,7 +22,7 @@ func TestApi_Image_Pull(t *testing.T) {
 	}
 
 	for _, testCase := range testCases {
-		id, err := api.ImagePull(ctx, testCase.Image)
+		id, err := api.ImagePull(ctx, testCase.Image, ImageAuthConfig{})
 		if testCase.Exists {
 			assert.NoError(t, err)
 			assert.NotEqual(t, "", id)